requirements.txt
streamlit==1.39.0
sentence-transformers==3.2.1
faiss-cpu==1.9.0
groq==0.11.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.32.3
numpy==1.26.4
tqdm==4.67.1